Fix Net::DNS 1.49 testsuite breakage
As of 1.49, apparently the previously readonly call
RR::OPT->option('foo') will, as a side effect, *create* the option
"foo" with an option-length of zero if it doesn't already exist.

One part of our testsuite driver used this call to check for the
existence of the "cookie" option while parsing packets for tests
which did not use the "cookie" option, resulting in failures from
the artificially-created "cookie" option.  This commit switches
the option() calls to use direct access to the option hash
instead, avoiding the side-effect.